+++ CHANGELOG.md
@@ -7,4 +7,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 ### Added
+- Elasticsearch(version 5.0.0) container is added.
 ### Fixed